working on UnitTests

+++ src/main/java/twistthrottle/services/impl/ProductServiceImpl.java
@@ -9,7 +9,6 @@ import twistthrottle.models.entities.enums.productType;
 import twistthrottle.repositories.ProductRepository;
 import twistthrottle.services.ProductService;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 @Service
@@ -51,7 +50,7 @@ public class ProductServiceImpl implements ProductService {
 
 
     @Override
-    public List<Product> findAllByPriceBetween(BigDecimal minPrice, BigDecimal maxPrice) {
+    public List<Product> findAllByPriceBetween(int minPrice, int maxPrice) {
         return productRepository.findAllByPriceBetween(minPrice,maxPrice);
     }
